+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix null in enrollment request without approval email
+- Fix implementation for user attribute update
 
 ## [1.0.0rc1] - 2024-10-25